auth-serrvice/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.0.0
sqlalchemy==2.0.22
slowapi==0.1.8
itsdangerous==2.1.2
python-dotenv==1.0.0
asyncpg==0.28.0
requests==2.31.0
redis==5.0.1
fastapi-cache2==0.2.1
prometheus-fastapi-instrumentator==6.1.0
